+++ src/pages/Home/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import styles from './Home.module.css'
+import styles from './index.module.css'
 import { Form, ArticleSection } from '../../common'
 import logo from '../../assets/logo.png'
