+++ src/pages/Team/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import styles from './Team.module.css'
+import styles from './index.module.css'
 import { ArticleSection, PersonCard } from '../../common';
 import penny from '../../assets/penny.jpg'
 import bryan from '../../assets/bryan.jpg'
