+++ src/pages/index.jsx
@@ -1,9 +1,9 @@
-export { default as About } from './About/About'
+export { default as About } from './About'
 export { default as BusinessSponsorship } from './BusinessSponsorship'
-export { default as Contact } from './Contact/Contact'
-export { default as Home } from './Home/Home'
-export { default as Message } from './Message/Message'
-export { default as SubmitMessage } from './SubmitMessage/SubmitMessage'
-export { default as Support } from './Support/Support'
-export { default as Team } from './Team/Team'
-export { default as TandC } from './TandC/TandC'
+export { default as Contact } from './Contact'
+export { default as Home } from './Home'
+export { default as Message } from './Message'
+export { default as SubmitMessage } from './SubmitMessage'
+export { default as Support } from './Support'
+export { default as Team } from './Team'
+export { default as TandC } from './TandC'
